updated printf print_address

+++ crss/printf/includes/ft_printf.h
@@ -31,7 +31,7 @@ int		ft_printf(char const *str, ...);
 //Number print functions
 size_t	ft_print_int(int nbr);
 size_t	ft_print_hexbyte(unsigned char byte, int is_big, char *flags);
-size_t	ft_print_hex(int hex, int is_big, char *flags);
+size_t	ft_print_hex(unsigned long long hex, int is_big, char *flags);
 size_t	ft_print_uint(unsigned int u_int);
 
 //Text print functions
